Find make from PATH when tests with `-v' flag fail
  Some platforms provide a working `make' program out of the box, but
some versions of this program will not return a successful exit status
code when given `-v' as argument.
  One such platform is FreeBSD, providing the `make' program in the base
system which is based on `pmake'. This version does not accept `-v':

$ make -v
usage: make [-BeikNnqrstWwX]
            [-C directory] [-D variable] [-d flags] [-f makefile]
            [-I directory] [-J private] [-j max_jobs] [-m directory] [-T file]
            [-V variable] [variable=value] [target ...]
zsh: exit 2     make -v

  This change ensures detection with `gmake -v` or `make -v` will
continue to work as expected, but adds a fall-back to return the first
executable entry named `make' which may be found in any of the
directories listed by the PATH environment variable.